support maxOccurs values + formats + adjust workflow to match generated WPS1-packages
