+++ profiles/doc.cfg
@@ -6,4 +6,4 @@ parts += sphinx
 recipe = zc.recipe.egg
 eggs =
     sphinx
-    ${twitcher:eggs}
+    ${weaver:eggs}
