+++ profiles/test.cfg
@@ -7,4 +7,4 @@ recipe = zc.recipe.egg
 eggs =
     pytest
     mock
-    ${twitcher:eggs}
+    ${weaver:eggs}
