+++ tests/functional/application-packages/Execute_ColibriFlyingpigeon_SubsetBbox.json
@@ -3,7 +3,7 @@
   "response": "document",
   "inputs": [
     {
-      "id": "tasmax",
+      "id": "resource",
       "href": "https://pavics.ouranos.ca/twitcher/ows/proxy/thredds/fileServer/birdhouse/nrcan/nrcan_canada_daily_v2/tasmax/nrcan_canada_daily_tasmax_2017.nc"
     },
     {
